use Rook's helm charts to deploy two Ceph clusters
This commit makes the e2e tests deploy two different Rook/Ceph clusters.
They will be used in the succeeding commits to check that multiple
mantle-controllers can coexist in the same k8s cluster.

In order to deploy two Rook/Ceph clusters easily, we use the Rook's helm
charts (`rook-ceph` and `rook-ceph-cluster`) in this commit, instead of
applying the bare manifests directly. `rook-ceph` is used to introduce a
rook-operator in the `rook-ceph` namespace, and `rook-ceph-cluster` is
used to deploy CephCluster resources in the `rook-ceph` and `rook-ceph2`
namespaces.