Avoid reconciling unrelated MantleBackup by checking its PVC's storageclass
Currently, the reconciler doesn't check that the PV(C) specified in a
given MantleBackup resource exists in the Rook/Ceph cluster that it's in
charge of.  Thus, under some circumstances, it may create a snapshot on
a wrong Rook/Ceph cluster.

This commit fixes the above problem. The reconciler now retrieves the
namespace where it's running, and assumes it to be the Ceph clusterID
that it's in charge of. If the ID doesn't match the clusterID specified
in the StorageClass of the MantleBackup's PV(C), the reconciler will
stop processing the MantleBackup resource.