get MantleBackup resource every time before updating its status
The current implementation updates MantleBackup's status without getting
its present content. Thus, if we try to update MantleBackup's status several
times in one reconciliation iteration, we fail to do so because its
resourceVersion field is invalid in the second update.

In order to resolve the problem above, this commit introduces a new
function named updateMantleBackupStatus. It not only gets the status
every time before updating it, but also allows us to update it in a
similiar manner to ctrl.CreateOrUpdate.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>