Use CephCmd in mantlebackup_controller
The previous code simulates the executeCommand function and sets static
timestamps for RBD snapshots. This approach poses a problem for
implementing expiration, as all MantleBackup resources would expire
during testing if the timestamps become outdated. To address this, we
need to create a more realistic fake CephCmd and use it.

Signed-off-by: Toshikuni Fukaya <toshikuni-fukaya@cybozu.co.jp>