fix mantlebackup_controller_test
We stopped setting SyncedToRemote=True in an ad hoc manner in the
previous commit, so we can't use `resMgr.WaitForbackupSyncedToRemote` to
wait for all necessary resources to be ready.

This commit resolves the problem above by packing all the tests into
`Eventually` in order to wait the resources to be ready.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>