test/e2e: use fixed names for StorageClass and CephBlockPool
We need to give a StorageClass name as a command-line argument of
mantle-controller so that it can create PVCs to store its exported data.
To do so, we need to fix the StorageClass name (and the pool name used
as a parameter of the StorageClass) before starting mantle-controller.
However, the current e2e test generates the names of the StorageClass
and pool in runtime, after starting the controller.

This commit resolves the problem above, by avoiding generating the names
in runtime and fixing them as constants ("rook-ceph-block") instead.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>