test/e2e: fix for async deletion of PVs and RBD images after MantleRestore removal
Some of the current e2e tests expect that, when a user deletes a
MantleRestore, mantle-controller tries to remove its corresponding PVs
and RBD images, and, if it can't do so, it hinders deletion of the
MantleRestore. However, this expectation isn't appropriate anymore,
because PVs and RBD images are now removed asynchronously.

This commit fixes the problem above by amending the tests.

First, the test "It should wait while the PV/PVC is in used" is amended
as follows:

- The title is changed to "It should NOT delete an RBD image while its
  corresponding PV and PVC are in use".
- It doesn't expect mantle-controller to prevent the ManlteRestore from
  being deleted anymore, even when a Pod is using the PV and PVC created by the
  MantleRestore. Instead, it does expect the MantleRestore to be removed,
  and the PV and the RBD image NOT to be removed.
- After the pod stops using the PV, the PV and the RBD image should be
  deleted.

Second, the test "It should fail to delete the MantleRestore if failed
to remove rbd image" is removed totally. This is because we couldn't
find any effective way to simulate a failure in deleting an RBD image.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>