check sync-mode annot is set before starting to reconcile MB in secondary
This commit ensures that sync-mode annotation is correctly set before
discard and import Jobs are created. Currently, mantle-controller does
NOT wait for a discard Job to finish before an import Job is created,
because an empty sync-mode passes this if condition:

https://github.com/cybozu-go/mantle/blob/fc095a2c6c9944faf2889aad703ee39e41ccc872/internal/controller/mantlebackup_controller.go#L1532-L1534

This behaviour isn't what we expect. This commit fixes this problem.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>