test/e2e: lenghthen timeout to wait for SyncToRemote to be True
According to some tries, the current timeout 5m is too short to wait for
SyncToRemote to be true in the e2e test, probably due to the exponential
backoff of `Requeue: true`. This commit lengthens its value to 10m to
resolve this problem in an ad hoc way. A fundamental resolution should
come in another PR.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>