test/e2e/multik8s: move tests in changetosecondary to replication package
The changetosecondary package was made to parallelize the e2e tests.
However, we can use labels for this purpose. This commit moves the tests
in the changetosecondary package to replication package. Instead, it
attaches "change-to-secondary" label to the tests so that they can run in
parallel.

There should be no functional changes in this commit.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>