test: controller: wait restoring PV creation before its deletion
The current implementation of `testDeleteResotringPV` is unstable and it
often fails with the following error message:

------------------------------
• [FAILED] [180.009 seconds]
MantleRestoreReconciler unit test test deleteRestoringPV [It] should delete the PV
/home/rbanno/mantle2/internal/controller/mantlerestore_controller_test.go:371

  [FAILED] Timed out after 180.000s.
  The function passed to Eventually failed at /home/rbanno/mantle2/internal/controller/mantlerestore_controller_test.go:389 with:
  Expected an error to have occurred.  Got:
      <nil>: nil
  In [It] at: /home/rbanno/mantle2/internal/controller/mantlerestore_controller_test.go:391 @ 12/09/25 02:34:26.574
------------------------------

It is probably because `test.reconciler.deleteRestoringPV` is called
before the restoring PV is actually created. To resolve this problem
this commit inserts code to wait for the PV to be created before its
deletion.

Signed-off-by: Ryotaro Banno <ryotaro.banno@gmail.com>