test: e2e: wait for MantleBackup to be deleted in DeleteMantleBackup
This is a _kaizen_ to stabilize the CI.

Currently, DeleteMantleBackup doesn't wait for the MantleBackup resource
to be deleted completely by using the `--wait=false` option. However,
some callers of this function expect the resource to be deleted
immediately after its call, so there is a race condition. The others,
which correctly consider the possibility, also wait for its deletion
after calling DeleteMantleBackup, so there's no need to skip waiting
anyway.

This commit removes `--wait=false` in DeleteMantleBackup and a function
that becomes unnecessary by this change. In addition, we tweak the
timeouts to make DeleteMantleBackup work correctly.